+++ Wrappers/Python/cil/processors/__init__.py
@@ -21,3 +21,4 @@ from .Normaliser import Normaliser
 from .Slicer import Slicer
 from .Binner import Binner
 from .RingRemover import RingRemover
+from .MaskGenerator import MaskGenerator